Added interface for MetricsConfiguration. Renamed existing MetricsConfiguration class to MetricsXMLConfiguration
